--- src/main/java/com/github/siboxd/fatturapa/model/CondizioniPagamentoType.java
@@ -1,14 +1,0 @@
-package com.github.siboxd.fatturapa.model;
-
-import org.simpleframework.xml.Namespace;
-import org.simpleframework.xml.Root;
-
-@Root(name = "CondizioniPagamentoType")
-@Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
-public enum CondizioniPagamentoType {
-
-    TP01,
-    TP02,
-    TP03
-
-}
