--- src/main/java/com/github/siboxd/fatturapa/model/DatiVeicoliType.java
@@ -1,40 +1,0 @@
-package com.github.siboxd.fatturapa.model;
-
-import org.simpleframework.xml.Element;
-import org.simpleframework.xml.Root;
-
-
-/**
- * DatiVeicoliType<br>
- * Generated using Android JAXB<br>
- *
- * @link https://github.com/yeshodhan/android-jaxb
- */
-@Root(name = "DatiVeicoliType")
-public class DatiVeicoliType {
-
-    @Element(name = "Data", required = true)
-    private String data;
-    @Element(name = "TotalePercorso", required = true)
-    private String totalePercorso;
-
-    public DatiVeicoliType() {
-    }
-
-    public String getData() {
-        return data;
-    }
-
-    public void setData(String data) {
-        this.data = data;
-    }
-
-    public String getTotalePercorso() {
-        return totalePercorso;
-    }
-
-    public void setTotalePercorso(String totalePercorso) {
-        this.totalePercorso = totalePercorso;
-    }
-
-}
