--- src/main/java/com/github/siboxd/fatturapa/model/ModalitaPagamentoType.java
@@ -1,33 +1,0 @@
-package com.github.siboxd.fatturapa.model;
-
-import org.simpleframework.xml.Namespace;
-import org.simpleframework.xml.Root;
-
-@Root(name = "ModalitaPagamentoType")
-@Namespace(reference = "http://ivaservizi.agenziaentrate.gov.it/docs/xsd/fatture/v1.2")
-public enum ModalitaPagamentoType {
-
-    MP01,
-    MP02,
-    MP03,
-    MP04,
-    MP05,
-    MP06,
-    MP07,
-    MP08,
-    MP09,
-    MP10,
-    MP11,
-    MP12,
-    MP13,
-    MP14,
-    MP15,
-    MP16,
-    MP17,
-    MP18,
-    MP19,
-    MP20,
-    MP21,
-    MP22
-
-}
